# backend/docs/swagger.yaml
openapi: 3.0.3
info:
  title: Hotel Booking System API
  description: |
    ระบบจองโรงแรมและที่พัก - RESTful API Documentation
    
    ## Overview
    ระบบนี้ถูกออกแบบมาเพื่อจัดการการจองโรงแรมแบบครบวงจร รองรับผู้ใช้งาน 4 บทบาท:
    - **Guest**: ผู้เข้าพัก - ค้นหาและจองห้องพัก
    - **Receptionist**: พนักงานต้อนรับ - จัดการ check-in/check-out
    - **Housekeeper**: แม่บ้าน - จัดการการทำความสะอาดห้อง
    - **Manager**: ผู้จัดการ - จัดการราคา สต็อก และรายงาน
    
    ## Authentication
    ระบบใช้ JWT (JSON Web Token) สำหรับการยืนยันตัวตน
    - ลงทะเบียนหรือเข้าสู่ระบบเพื่อรับ access token
    - ใส่ token ใน Authorization header: `Bearer <token>`
    
    ## Rate Limiting
    - Auth endpoints: 5 requests/minute
    - Search endpoints: 20 requests/minute
    - Booking endpoints: 10 requests/minute
    - General endpoints: 100 requests/minute
    
  version: 1.0.0
  contact:
    name: Hotel Booking System Support
    email: support@hotelbooking.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Development server
  - url: http://localhost:8080/api
    description: Development API server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Error message"
...
    RegisterRequest:
      type: object
      required:
        - email
        - password
        - firstName
        - lastName
        - phone
      properties:
        email:
          type: string
          format: email
          example: "guest@example.com"
        password:
          type: string
          format: password
          minLength: 8
          example: "SecurePass123!"
        firstName:
          type: string
          example: "สมชาย"
        lastName:
          type: string
          example: "ใจดี"
        phone:
          type: string
          example: "0812345678"
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "guest@example.com"
        password:
          type: string
          format: password
          example: "SecurePass123!"
    LoginResponse:
      type: object
      properties:
        guestId:
          type: integer
          example: 1
        email:
          type: string
          example: "guest@example.com"
        firstName:
          type: string
          example: "สมชาย"
        lastName:
          type: string
          example: "ใจดี"
        role:
          type: string
          enum: [guest, receptionist, housekeeper, manager]
          example: "guest"
        accessToken:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
    Guest:
      type: object
      properties:
        guestId:
          type: integer
          example: 1
        firstName:
          type: string
          example: "สมชาย"
        lastName:
          type: string
          example: "ใจดี"
        email:
          type: string
          example: "guest@example.com"
        phone:
          type: string
          example: "0812345678"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    RoomType:
      type: object
      properties:
        roomTypeId:
          type: integer
          example: 1
        name:
          type: string
          example: "Deluxe Room"
        description:
          type: string
          example: "ห้องพักหรูพร้อมวิวทะเล"
        maxOccupancy:
          type: integer
          example: 2
        defaultAllotment:
          type: integer
          example: 10
        amenities:
          type: array
          items:
            $ref: '#/components/schemas/Amenity'
        images:
          type: array
          items:
            type: string
          example: ["/images/deluxe-1.jpg", "/images/deluxe-2.jpg"]
        basePrice:
          type: number
          format: decimal
          example: 2500.00
    Amenity:
      type: object
      properties:
        amenityId:
          type: integer
          example: 1
        name:
          type: string
          example: "WiFi"
...
    SearchRoomsResponse:
      type: object
      properties:
        availableRooms:
          type: array
          items:
            $ref: '#/components/schemas/RoomType'
        checkIn:
          type: string
          format: date
        checkOut:
          type: string
          format: date
        nights:
          type: integer
          example: 4
    
    # Booking Models
...
paths:
...
  /api/auth/register:
    post:
      tags:
        - auth
      summary: Register a new guest
      description: Create a new guest account with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              email: "newguest@example.com"
              password: "SecurePass123!"
              firstName: "สมชาย"
              lastName: "ใจดี"
              phone: "0812345678"
      responses:
        '201':
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          description: Validation error or email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                emailExists:
                  value:
                    error: "อีเมลนี้ถูกลงทะเบียนแล้ว"
                validationError:
                  value:
                    error: "Invalid email format"
  /api/auth/login:
    post:
      tags:
        - auth
      summary: Login
      description: Authenticate with email and password to receive JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: "guest@example.com"
              password: "SecurePass123!"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: "อีเมลหรือรหัสผ่านไม่ถูกต้อง"
  /api/auth/me:
    get:
      tags:
        - auth
      summary: Get current user profile
      description: Retrieve the authenticated user's profile information
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Guest'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Guest not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/rooms/search:
    get:
      tags:
        - rooms
      summary: Search available rooms
      description: Search for available rooms based on check-in/check-out dates and number of guests
      parameters:
        - name: checkIn
          in: query
          required: true
          schema:
            type: string
            format: date
          example: "2025-12-01"
        - name: checkOut
          in: query
          required: true
          schema:
            type: string
            format: date
          example: "2025-12-05"
        - name: guests
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
          example: 2
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchRoomsResponse'
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                invalidDate:
                  value:
                    error: "รูปแบบวันที่ check-in ไม่ถูกต้อง"
                pastDate:
                  value:
                    error: "วันที่ check-in ต้องไม่อยู่ในอดีต"
...
  /api/rooms/types/{id}:
    get:
      tags:
        - rooms
      summary: Get room type by ID
      description: Retrieve detailed information about a specific room type
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        '200':
          description: Room type details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomType'
        '404':
          description: Room type not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/rooms/types/{id}/pricing:
    get:
      tags:
        - rooms
      summary: Get room type pricing
      description: Calculate pricing for a specific room type for given dates
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
        - name: checkIn
          in: query
          required: true
          schema:
            type: string
            format: date
          example: "2025-12-01"
        - name: checkOut
          in: query
          required: true
          schema:
            type: string
            format: date
          example: "2025-12-05"
      responses:
        '200':
          description: Pricing information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomType'
        '404':
          description: Room type not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/rooms/status:
    get:
      tags:
        - rooms
      summary: Get room status dashboard
      description: Get all rooms with their current occupancy and housekeeping status (Receptionist only)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Room status dashboard
          content:
            application/json:
              schema:
                type: object
                properties:
                  rooms:
                    type: array
                    items:
                      type: object
                      properties:
                        roomId:
                          type: integer
                        roomNumber:
                          type: string
                        roomTypeName:
                          type: string
                        occupancyStatus:
                          type: string
                        housekeepingStatus:
                          type: string
                        guestName:
                          type: string
                          nullable: true
                        checkOutDate:
                          type: string
                          format: date
                          nullable: true
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - Receptionist role required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Booking Endpoints
...